+++ xtask/src/main.rs
@@ -166,7 +166,8 @@ impl CodemirrorVendor {
         );
         index_js.push_str("export { EditorView, minimalSetup } from \"./codemirror.js\";\n");
         index_js.push_str(
-            "export { EditorState, EditorSelection, Annotation } from \"./codemirror__state.js\";\n",
+            "export { EditorState, EditorSelection, Annotation, Prec } \
+             from \"./codemirror__state.js\";\n",
         );
         index_js.push_str(
             "export { lineNumbers, highlightActiveLineGutter, highlightActiveLine, \
@@ -175,14 +176,19 @@ impl CodemirrorVendor {
              from \"./codemirror__view.js\";\n",
         );
         index_js.push_str(
-            "export { HighlightStyle, syntaxHighlighting, bracketMatching, indentOnInput } \
+            "export { HighlightStyle, syntaxHighlighting, bracketMatching, indentOnInput, \
+             foldGutter, foldKeymap, codeFolding, foldable, foldEffect, unfoldEffect, \
+             foldedRanges, unfoldAll, indentUnit } \
              from \"./codemirror__language.js\";\n",
         );
         index_js.push_str("export { SearchCursor } from \"./codemirror__search.js\";\n");
         index_js.push_str(
             "export { closeBrackets, closeBracketsKeymap } from \"./codemirror__autocomplete.js\";\n",
         );
-        index_js.push_str("export { indentWithTab } from \"./codemirror__commands.js\";\n");
+        index_js.push_str(
+            "export { indentWithTab, moveLineUp, moveLineDown } \
+             from \"./codemirror__commands.js\";\n",
+        );
         index_js.push_str("export { tags } from \"./lezer__highlight.js\";\n");
         index_js.push_str(
             "export { LSPClient, LSPPlugin, languageServerExtensions } \