+++ CHANGELOG.md
@@ -1,12 +1,24 @@
 # Changelog
 
+## unreleased
+
+* Support code folding. ([#4][#4])
+
+[#4]: https://github.com/azriel91/dioxus_codemirror/pull/4
+
+
 ## 0.2.0 (2026-06-07)
 
-* Support code actions with `Ctrl/Cmd + .` keybinding.
-* Support theme colour overrides via the `theme_colors` prop (`ThemeColors` / `ThemeColor`).
+* Support code actions with `Ctrl/Cmd + .` keybinding. ([#2][#2])
+* Support theme colour overrides via the `theme_colors` prop (`ThemeColors` / `ThemeColor`). ([#3][#3])
+
+[#2]: https://github.com/azriel91/dioxus_codemirror/pull/2
+[#3]: https://github.com/azriel91/dioxus_codemirror/pull/3
 
 
 ## 0.1.0 (2026-06-06)
 
-* Add `CodeMirror` dioxus component with CodeMirror 6 integration.
-* Support some CodeMirror features.
+* Add `CodeMirror` dioxus component with CodeMirror 6 integration. ([#1][#1])
+* Support some CodeMirror features. ([#1][#1])
+
+[#1]: https://github.com/azriel91/dioxus_codemirror/pull/1